COM-1164: Add support for importing multiselect enum values (#106)
## Description

Previously only string values could be imported in the contact import.
In case the project uses multiselect for certain contact attributes the
import was not working correctly and was throwing a validation error
since the attribute was a string but was expecting an array. Now the
contact attribute gets checked if it is supposed to be an array and gets
transformed to an array. The values need to be separated with a comma in
the import file.

## Example

[x] I have verified if my change requires an example


## Changeset

[x] I have verified if my change requires a changeset

## Related tasks and documents

COM-1164

---------

Co-authored-by: Denise Buder <denise.buder@vivid-planet.com>